Pass redirect path in a cookie
