+++ config/default.yaml
@@ -38,7 +38,7 @@ services:
     microservice: 'divorce_frontend'
     secret: 'NEED TO INSERT SECRET'
   idam:
-    authorizationURL: 'https://idam-web-public.aat.platform.hmcts.net/login'
+    authorizationURL: 'https://idam-web-public.aat.platform.hmcts.net/o/authorize'
     tokenURL: 'https://idam-api.aat.platform.hmcts.net/o/token'
     endSessionURL: 'https://idam-web-public.aat.platform.hmcts.net/o/endSession'
     authorizationScope: 'openid profile roles'
