+++ src/main/steps/applicant1/check-your-answers/joint-form.njk
@@ -22,7 +22,7 @@
     {% block after_form %}{% endblock %}
 
     <p class="govuk-body">
-      <button class="hmcts-button-link" type="submit" name="saveAndSignOut" value="true" data-prevent-double-click="true" data-module="govuk-button" />
+      <button class="hmcts-button-link" type="submit" name="saveAndSignOut" value="true" data-prevent-double-click="true" data-module="govuk-button">
         {{ saveAndSignOut }}
       </button>
