+++ src/main/steps/applicant1/withdraw-application/post.ts
@@ -5,7 +5,7 @@ import { Response } from 'express';
 import { CITIZEN_WITHDRAWN } from '../../../app/case/definition';
 import { AppRequest } from '../../../app/controller/AppRequest';
 import { AnyObject, PostController } from '../../../app/controller/PostController';
-import { destroySessionAndRedirectToSignOut } from '../../../app/controller/controller.utils';
+import { destroySessionAndRedirectToSignOut } from '../../../app/controller/signout';
 import { APPLICATION_WITHDRAWN } from '../../urls';
 
 const logger = Logger.getLogger('withdraw-application-controller');
