+++ src/main/steps/applicant1/you-can-use-english-welsh-courts/template.njk
@@ -62,7 +62,7 @@
       </form>
 
       <p class="govuk-body">
-        <button class="hmcts-button-link" type="submit" name="saveAndSignOut" value="true" data-prevent-double-click="true" data-module="govuk-button" />
+        <button class="hmcts-button-link" type="submit" name="saveAndSignOut" value="true" data-prevent-double-click="true" data-module="govuk-button">
           {{ saveAndSignOut }}
         </button>
       </p>
