+++ src/main/steps/request-for-information-save-sign-out/get.ts
@@ -3,7 +3,7 @@ import { Response } from 'express';
 
 import { AppRequest } from '../../app/controller/AppRequest';
 import { GetController } from '../../app/controller/GetController';
-import { destroySessionAndRedirectToSignOutViaCallback } from '../../app/controller/controller.utils';
+import { destroySessionAndRedirectToSignOutViaCallback } from '../../app/controller/signout';
 
 import { generateContent } from './content';
 import { REQUEST_FOR_INFORMATION_SAVE_AND_SIGN_OUT } from '../urls';
